Refactor teacher availability feature to use GraphQL
Working on teacher availability feature to utilize GraphQL queries and mutations instead of rest APIs. Added ApolloClient to frontend and setup proper error handling for production and development environments in the backend. Included new HttpExceptionFilter in backend for handling exceptions and generated new GraphQL queries and mutations in separate files.